Refactor some stuff and add some support that may be used for en passant
